blob_upload_appservice/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
azure-cosmos==4.9.0
azure-storage-blob
